# api/openapi.yaml
openapi: 3.1.0
info:
  title: Ant Colony Manager API
  version: "1"
  description: |
    Self-hosted API for managing ant colonies. Used by the Android app and the web app.

    **Conventions**
    - All IDs are UUIDs (UUIDv7, usually generated by the client).
    - Timestamps are RFC 3339.
    - Errors use `application/problem+json` (RFC 9457) with a machine-readable `code`.
    - Resources of colonies you have no access to always answer **404** (existence is never revealed).
    - Every write (REST or sync) is idempotent: REST writes accept `Idempotency-Key` (UUID),
      sync operations carry an `op_id`.
    - Optimistic concurrency: send `If-Match: "<version>"` on PATCH; conflicting fields are merged
      (see `WriteResponse.conflicts`).
servers:
  - url: "{publicUrl}"
    variables:
      publicUrl:
        default: https://ants.example.com
security:
  - bearer: []
...
paths:
...
  /c/{token}:
    get:
      tags: [scan]
      security: []
      summary: "Scan landing page (QR/NFC without app). Contains no colony data."
      parameters: [{ $ref: "#/components/parameters/Token" }]
      responses: { "200": { description: "HTML" } }
...
  /api/v1/setup:
    post:
      tags: [auth]
      security: []
      summary: "Create the first administrator (requires the setup token from the server log)"
      requestBody: { $ref: "#/components/requestBodies/Register" }
      responses:
        "201": { $ref: "#/components/responses/Session" }
        "403": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
  /api/v1/auth/register:
    post:
      tags: [auth]
      security: []
      summary: "Register (depending on REGISTRATION_MODE; invite_token may also share a colony)"
      requestBody: { $ref: "#/components/requestBodies/Register" }
      responses:
        "201": { $ref: "#/components/responses/Session" }
        "403": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /api/v1/auth/login:
    post:
      tags: [auth]
      security: []
      summary: "Log in. Web clients send `X-ACM-Client: web` and receive the refresh token as HttpOnly cookie."
      parameters: [{ $ref: "#/components/parameters/ClientHeader" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
                device: { $ref: "#/components/schemas/Device" }
      responses:
        "200": { $ref: "#/components/responses/Session" }
        "401": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/Problem" }
  /api/v1/auth/refresh:
    post:
      tags: [auth]
      security: []
      summary: "Rotate the refresh token (body for apps, cookie + X-ACM-Client for web)"
      description: |
        Reusing an already rotated token after 30 s revokes the whole session family
        (`auth.token_reused`). Within 30 s the answer is `auth.token_rotated` (parallel refresh).
      parameters: [{ $ref: "#/components/parameters/ClientHeader" }]
      requestBody:
        content:
          application/json:
            schema: { type: object, properties: { refresh_token: { type: string } } }
      responses:
        "200": { $ref: "#/components/responses/Session" }
        "401": { $ref: "#/components/responses/Problem" }
...
  /api/v1/auth/password/forgot:
    post:
      tags: [auth]
      security: []
      summary: "Send a reset e-mail. Always 202 (no account enumeration)."
      requestBody:
        content: { application/json: { schema: { type: object, properties: { email: { type: string } } } } }
      responses: { "202": { description: "accepted" } }
  /api/v1/auth/password/reset:
    post:
      tags: [auth]
      security: []
      summary: "Set a new password with a reset token (revokes all sessions)"
      requestBody:
        content:
          application/json:
            schema: { type: object, required: [token, password], properties: { token: { type: string }, password: { type: string } } }
      responses:
        "204": { description: "done" }
        "422": { $ref: "#/components/responses/Problem" }
...
  /api/v1/auth/device-link/redeem:
    post:
      tags: [auth]
      security: []
      summary: "App exchanges the code for a session"
      requestBody:
        content:
          application/json:
            schema: { type: object, required: [code], properties: { code: { type: string }, device: { $ref: "#/components/schemas/Device" } } }
      responses:
        "200": { $ref: "#/components/responses/Session" }
        "422": { $ref: "#/components/responses/Problem" }
...
  /api/v1/auth/sessions/{id}:
    delete:
      tags: [auth]
      summary: "Sign out a device (the app wipes local data on next contact)"
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses: { "204": { description: "revoked" } }
...
  /api/v1/sync/push:
    post:
      tags: [sync]
      summary: "Apply up to 100 operations in order, each exactly once"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/Device"
                - type: object
                  required: [ops]
                  properties:
                    ops: { type: array, maxItems: 100, items: { $ref: "#/components/schemas/Op" } }
      responses:
        "200":
          description: "results in request order"
          content:
            application/json:
              schema:
                type: object
                properties:
                  results: { type: array, items: { $ref: "#/components/schemas/OpResult" } }
                  server_seq: { type: integer, format: int64 }
  /api/v1/sync/pull:
    get:
      tags: [sync]
      summary: "Changes after a cursor (including tombstones)"
      parameters:
        - { name: since, in: query, schema: { type: integer, format: int64, default: 0 } }
        - { name: limit, in: query, schema: { type: integer, default: 500, maximum: 1000 } }
        - { name: X-Device-ID, in: header, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: "changes"
          content:
            application/json:
              schema:
                type: object
                properties:
                  changes: { type: array, items: { $ref: "#/components/schemas/Change" } }
                  next: { type: integer, format: int64, description: "cursor for the next pull" }
                  has_more: { type: boolean }
        "410":
          description: "cursor older than the tombstone horizon → run a snapshot (`sync.resync_required`)"
          content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
  /api/v1/sync/snapshot:
    get:
      tags: [sync]
      summary: "Full current state (first sync, resync, or newly shared colonies via colony_ids)"
      parameters:
        - { name: colony_ids, in: query, schema: { type: string }, description: "comma-separated UUIDs" }
      responses:
        "200":
          description: "snapshot"
          content:
            application/json:
              schema:
                type: object
                properties:
                  cursor: { type: integer, format: int64 }
                  entities: { type: object, additionalProperties: { type: array, items: { type: object } } }
...
  /api/v1/sync/conflicts/{id}:
    delete:
      tags: [sync]
      summary: "Dismiss a conflict hint"
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses: { "204": { description: "dismissed" } }
...
  /api/v1/export.zip:
    get:
      tags: [colonies]
      summary: "ZIP: export.json, CSV tables (semicolon, UTF-8 BOM) and all photos"
      parameters:
        - { name: photos, in: query, schema: { type: string, enum: ["0", "1"] }, description: "0 = without photos" }
      responses: { "200": { description: "file download (application/zip)" } }
...
  /api/v1/me/notifications:
    get:
      tags: [auth]
      summary: "Notification settings: ntfy address, channels per topic (digest, overdue, sensor, winter), repeat, quiet hours. The ntfy token is never returned (ntfy_token_set)"
      responses: { "200": { description: "settings" } }
    put:
      tags: [auth]
      summary: "Replace notification settings; ntfy_token omitted = keep, \"\" = remove. Repeat hours: overdue 0/6/12/24, sensor 0/1/6/12/24, winter 0/24 (0 = only once)"
      responses: { "200": { description: "saved settings" }, "422": { description: "invalid value" } }
  /api/v1/admin/smtp:
    get:
      tags: [admin]
      summary: "E-mail server settings (admin). source: app | env | none; the password is never returned (password_set)"
      responses: { "200": { description: "settings" }, "403": { description: "not an administrator" } }
    put:
      tags: [admin]
      summary: "Save SMTP settings (take effect immediately, precedence over SMTP_*); host \"\" removes them; password omitted = keep, \"\" = remove"
      responses: { "200": { description: "saved settings" }, "422": { description: "invalid value" } }
  /api/v1/admin/smtp/test:
    post:
      tags: [admin]
      summary: "Send a test e-mail to the administrator"
      responses: { "204": { description: "sent" }, "502": { description: "SMTP failed (title explains)" } }
  /api/v1/admin/offsite:
    get:
      tags: [admin]
      summary: "Off-site backup settings and status {enabled, type (webdav | smb | folder), url (WebDAV address, smb://server/share/folder or a folder in the container), user, password_set, keep, status{last_name, last_success, last_error, last_error_at, last_bytes, last_photos, running, local_latest}}"
      responses: { "200": { description: "settings" }, "403": { description: "not an administrator" } }
    put:
      tags: [admin]
      summary: "Save off-site backup settings; password omitted = keep, \"\" = remove; keep = remote backups to keep (1–365)"
      responses: { "200": { description: "saved settings" }, "422": { description: "invalid value" } }
  /api/v1/admin/offsite/test:
    post:
      tags: [admin]
      summary: "Check the target: address, login and write access (creates the folder if missing; a mounted folder must exist)"
      responses: { "204": { description: "reachable" }, "502": { description: "target failed (title explains)" } }
  /api/v1/admin/offsite/run:
    post:
      tags: [admin]
      summary: "Upload the newest local backup now (in the background; status via GET /admin/offsite)"
      responses: { "202": { description: "started" }, "422": { description: "not set up" } }
  /api/v1/admin/mqtt:
    get:
      tags: [admin]
      summary: "Home Assistant via MQTT: broker settings and status {enabled, url, user, password_set, prefix, owner, status{connected, last_sync, last_error, last_error_at, colonies}}"
      responses: { "200": { description: "settings (also members: other users who send their colonies; ha_url, ha_token_set: Home Assistant REST API for sensor values)" }, "403": { description: "not an administrator" } }
    put:
      tags: [admin]
      summary: "Save the MQTT broker (and optionally ha_url + ha_token for sensor values; ha_token omitted = keep, \"\" = remove) (url: host, host:port or mqtt(s)://host:port; prefix: discovery prefix, default homeassistant); password omitted = keep, \"\" = remove. From now on the colonies of this administrator are announced to Home Assistant (MQTT discovery); switching off removes them there"
      responses: { "200": { description: "saved settings" }, "422": { description: "invalid value" } }
  /api/v1/admin/update:
    get:
      tags: [admin]
      summary: "Update button: {available (updater service runs), state idle|requested|running|done|failed, message, at, log (after a failure)}"
      responses: { "200": { description: "status" }, "403": { description: "not an administrator" } }
    post:
      tags: [admin]
      summary: "Update now: the optional updater service backs up, pulls the images and restarts (like scripts/update.sh)"
      responses: { "202": { description: "requested" }, "409": { description: "no updater or already running" } }
  /api/v1/admin/ai:
    get:
      tags: [admin]
      summary: "Counting ants with AI: {enabled, provider (anthropic | openai | openrouter), api_key_set, model}"
      responses: { "200": { description: "settings" }, "403": { description: "not an administrator" } }
    put:
      tags: [admin]
      summary: "Save {enabled, provider, api_key (omitted = keep, \"\" = remove; changing the provider drops the old key), model (anthropic: default claude-opus-5-5; openai/openrouter: required, a model that understands images)}"
      responses: { "200": { description: "settings" }, "422": { description: "invalid" } }
...
  /api/v1/colonies/{id}/ai-count:
    post:
      tags: [colonies]
      summary: "Count the ants on 1–6 uploaded photos of the colony with AI {photo_ids, async?}. With async: true 202 {id, state} at once – poll /ai-count/{job}; otherwise waits for {photos[{photo_id, count, min, max, queens, note}], total, min, max, overlap, note, model} – the photos are sent to the provider chosen by the administrator"
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses: { "200": { description: "counts" }, "409": { description: "not set up" }, "422": { description: "invalid photos" }, "502": { description: "AI failed (title explains)" } }
  /api/v1/ai-count/{job}:
    get:
      tags: [colonies]
      summary: "State of a counting job started with {async: true}: {id, state running|done|failed, result?, error?} – only for the user who started it; kept 15 minutes"
      parameters:
        - { name: job, in: path, required: true, schema: { type: string, format: uuid } }
      responses: { "200": { description: "job" }, "404": { description: "unknown job" } }
  /api/v1/care-covers:
    get:
      tags: [colonies]
      summary: "Care covers (Pflegevertretung) I set up and those I stand in for: [{id, owner_name, user_name, user_email (owner only), starts_on, ends_on, instructions, state planned|active|ended, mine, colonies[{colony_id, name, number, instructions}]}]"
      responses: { "200": { description: "list" } }
    post:
      tags: [colonies]
      summary: "Hand own colonies to a person with an account for a period {email, starts_on, ends_on, instructions?, colonies[{colony_id, instructions?}]}; during it the person is a carer (editor)"
      responses: { "201": { description: "cover" }, "422": { description: "invalid (unknown e-mail, dates, foreign colony)" } }
  /api/v1/care-covers/{id}:
    get:
      tags: [colonies]
      summary: "One cover with done[{colony_id, type, occurred_at, note}] – the care the substitute documented during it"
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses: { "200": { description: "cover" }, "404": { description: "not found" } }
    patch:
      tags: [colonies]
      summary: "Change starts_on, ends_on, instructions or colonies (owner, not after the end)"
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses: { "200": { description: "cover" }, "404": { description: "not found" }, "422": { description: "invalid" } }
  /api/v1/care-covers/{id}/end:
    post:
      tags: [colonies]
      summary: "End a cover now (owner or substitute); a cover that has not started is removed"
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses: { "204": { description: "ended" }, "404": { description: "not found" } }
  /api/v1/colonies/{id}/care-instructions:
    get:
      tags: [colonies]
      summary: "Running and planned covers of this colony that concern me, with the general and this colony's instructions"
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses: { "200": { description: "list" }, "404": { description: "colony not found" } }
  /api/v1/admin/home-assistant/test:
    post:
      tags: [admin]
      summary: "Check the saved Home Assistant address and access token (REST API, for sensor values)"
      responses: { "204": { description: "reachable" }, "502": { description: "Home Assistant failed (title explains)" }, "422": { description: "not set up" } }
...
  /api/v1/admin/mqtt/test:
    post:
      tags: [admin]
      summary: "Connect to the saved MQTT broker once (address and login)"
      responses: { "204": { description: "reachable" }, "502": { description: "MQTT failed (title explains)" }, "422": { description: "not set up" } }
  /api/v1/snooze:
    post:
      tags: [colonies]
      security: []
      summary: "„Morgen“ button of an ntfy notification: signed link (7 days). k=care → care of the colony due by today counts as due tomorrow; k=winter → planned winter start/end +1 day"
      responses: { "200": { description: "{message}" }, "401": { description: "link invalid or expired" } }
  /api/v1/me/notifications/test:
    post:
      tags: [auth]
      summary: "Send a test message via the saved ntfy address"
      responses: { "204": { description: "sent" }, "502": { description: "ntfy refused or unreachable (title explains)" } }
...
  /api/v1/colonies/{id}/archive:
    post: { tags: [colonies], summary: "Archive (owner only)", parameters: [{ $ref: "#/components/parameters/Id" }], responses: { "200": { $ref: "#/components/responses/Write" } } }
...
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer, bearerFormat: JWT, description: "Access token (15 min)" }
    sensorKey: { type: http, scheme: bearer, description: "Sensor key acm_sk_<prefix>_<secret>" }
  parameters:
...
    Token: { name: token, in: path, required: true, schema: { type: string, pattern: "^[0-9A-Za-z]{16}$" } }
...
    ClientHeader: { name: X-ACM-Client, in: header, schema: { enum: [web] } }
...
  responses:
    Problem:
      description: "error"
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
    Session:
      description: "tokens and user"
      content: { application/json: { schema: { $ref: "#/components/schemas/Session" } } }
...